Alter Response Times alert to use histogram bucket
The current response times alert is triggered by changes to the average
response time, this can miss an increase in slow running requests if
that is masked by high volumes of quick requests.  The histogram bucket
looks at slowest performance for 95% of requests so provides a wider
view whilst still excluding outliers.